Use e2e-specific authflow client
