+++ e2e/tests/authflow/client/client.go
@@ -1,4 +1,4 @@
-package authflowclient
+package e2eclient
 
 import (
 	"bytes"
