+++ e2e/tests/receipe_1/expire_user.sql
@@ -5,9 +5,9 @@ with a as (
   JOIN _auth_identity ON _auth_authenticator.user_id = _auth_user.id
   JOIN _auth_identity_login_id ON _auth_identity_login_id.id = _auth_identity.id
 )
-update _auth_authenticator
-set updated_at = '2000-01-01 10:10:10'
-from a
+UPDATE _auth_authenticator
+SET updated_at = '2000-01-01 10:10:10'
+FROM a
 WHERE a.login_id = 'e2e_recipe_1_expiry@authgear.com'
 AND id = a.authenticator_id
 AND app_id = '{{ .AppID }}';
