--- pkg/lib/authflowclient/deps.go
@@ -1,9 +1,0 @@
-package authflowclient
-
-import (
-	"github.com/google/wire"
-)
-
-var DependencySet = wire.NewSet(
-	NewClient,
-)
