--- pkg/lib/authflowclient/errors.go
@@ -1,7 +1,0 @@
-package authflowclient
-
-import (
-	"github.com/authgear/authgear-server/pkg/api/apierrors"
-)
-
-var ErrFlowNotFound = apierrors.NotFound.WithReason("AuthenticationFlowNotFound").New("flow not found")
